feat: Improve profile empty state design
